tests/requirements.txt
pytest==4.0.2
pytest-cov==2.6.0
pytest-flakes==4.0.0
freezegun-0.3.11
